Изменен некорректный путь к файлу в классе ServiceReader Тип Double заменен на BigDecimal Добавлен функционал согласно заданию: подсчет минимальной прибыли по товару и услуге, сверка прибыли по товарам и услугам

+++ src/main/java/ru/leroymerlin/data/Revenue.java
@@ -2,11 +2,13 @@ package ru.leroymerlin.data;
 
 import ru.leroymerlin.protocol.ReportType;
 
+import java.math.BigDecimal;
+
 public class Revenue {
     private final ReportType reportType;
-    private final Double profit;
+    private final BigDecimal profit;
 
-    public Revenue(ReportType reportType, Double profit) {
+    public Revenue(ReportType reportType, BigDecimal profit) {
         this.reportType = reportType;
         this.profit = profit;
     }
@@ -24,7 +26,7 @@ public class Revenue {
         return reportType;
     }
 
-    public Double getProfit() {
+    public BigDecimal getProfit() {
         return profit;
     }
     //endregion
